web: reorganized folder structure and updated product display
