+++ src/Web/eShop.Web/src/app/app.routes.ts
@@ -1,8 +1,8 @@
 import { Routes } from '@angular/router';
 import { HeaderComponent } from './shared/header/header.component';
-import { LandingPageComponent } from './components/landing-page/landing-page.component';
-import { LoginPageComponent } from './components/login-page/login-page.component';
-import { SigninPageComponent } from './components/signin-page/signin-page.component';
+import { LandingPageComponent } from './components/home/landing-page/landing-page.component';
+import { LoginPageComponent } from './components/authentication/login-page/login-page.component';
+import { SigninPageComponent } from './components/authentication/signin-page/signin-page.component';
 import { CartPageComponent } from './components/cart-page/cart-page.component';
 import { ProductPageComponent } from './components/product-page/product-page.component';
 import { PaymentPageComponent } from './components/payment-page/payment-page.component';
