+++ src/Web/eShop.Web/src/app/components/home/landing-page/landing-page.component.ts
@@ -1,8 +1,8 @@
 import { Component, OnInit } from '@angular/core';
-import { HeaderComponent } from '../../shared/header/header.component';
+import { HeaderComponent } from '../../../shared/header/header.component';
 import { StaticCarouselComponent } from '../static-carousel/static-carousel.component';
 import { ProductDisplayComponent } from '../product-display/product-display.component';
-import { FooterComponent } from "../../shared/footer/footer.component";
+import { FooterComponent } from "../../../shared/footer/footer.component";
 import { CategoryDisplayComponent } from '../category-display/category-display.component';
 import { ViewportScroller } from '@angular/common';
 import { ActivatedRoute } from '@angular/router';
