+++ src/Web/eShop.Web/src/app/components/home/product-display/product-display.component.ts
@@ -1,7 +1,7 @@
 import { Router } from '@angular/router';
-import { Product } from './../../models/product.model';
-import { ProductService } from './../../services/product-list/product.service';
-import { Component, OnInit } from '@angular/core';
+import { Product } from '../../../models/product.model';
+import { ProductService } from '../../../services/product-list/product.service';
+import { Component, Input, OnInit } from '@angular/core';
 
 @Component({
   selector: 'app-product-display',
@@ -12,6 +12,8 @@ import { Component, OnInit } from '@angular/core';
 })
 export class ProductDisplayComponent implements OnInit {
 
+  @Input() category?: string
+
   products? : Product[]
 
   constructor(private router: Router, private productService: ProductService) { }
@@ -20,7 +22,7 @@ export class ProductDisplayComponent implements OnInit {
     this.productService.getCatalogItems()
       .subscribe({
         next : (response) => {
-          this.products = response
+          this.products = response.items
         }
       })
   }
