+++ src/Web/eShop.Web/src/app/components/home/static-carousel/static-carousel.component.spec.ts
