+++ src/Web/eShop.Web/src/app/models/product-request.model.ts
@@ -1,0 +1,8 @@
+import { Product } from "./product.model";
+
+export interface ProductRequest {
+  pageSize: number,
+  pageIndex: number,
+  count: number,
+  items: Product[]
+}
