+++ src/Web/eShop.Web/src/app/services/product-list/product.service.ts
@@ -2,6 +2,7 @@ import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { Observable } from 'rxjs';
 import { Product } from '../../models/product.model';
+import { ProductRequest } from '../../models/product-request.model';
 
 @Injectable({
   providedIn: 'root'
@@ -11,8 +12,8 @@ export class ProductService {
 
   constructor(private http: HttpClient) { }
 
-  getCatalogItems(): Observable<Product[]> {
-    return this.http.get<Product[]>(`${this.url}/items`)
+  getCatalogItems(): Observable<ProductRequest> {
+    return this.http.get<ProductRequest>(`${this.url}/items`)
   }
 
   getCalalogItem(id: string) : Observable<Product> {
